Paragraph: Only indent sub-block when it can't be reformatted
In other words, if a parapgraph contains another paragraph (e.g. to set
colors on part of the text), concatenate both as if they were a single
paragraph, because the sub-paragraph can be reformatted.

If instead a paragraph contains a table for instance, keep the table
format as it is and indent following lines to keep alignment.